Some dependency updates + remove gecko projectile setup

+++ install.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-cp ~/.emacs ~/.emacs.bak.$(date +%d%m%y-%H%m%S)
+if [[ -e ~/.emacs ]]; then
+    cp ~/.emacs ~/.emacs.bak.$(date +%d%m%y-%H%m%S)
+fi
 cp .emacs ~/.emacs
